Fix an issue where subpriority paging could be truncated
Ref T7548. Subpriority is a float, but we're truncating it to an int, which can cause reselection of the same row while paging.

+++ src/applications/maniphest/query/ManiphestTaskQuery.php
@@ -1066,8 +1066,8 @@ final class ManiphestTaskQuery extends PhabricatorCursorPagedPolicyAwareQuery {
           }
           $columns[] = array(
             'name' => 'task.subpriority',
-            'value' => (int)$cursor->getSubpriority(),
-            'type' => 'int',
+            'value' => $cursor->getSubpriority(),
+            'type' => 'float',
             'reverse' => true,
           );
           $columns[] = array(
